+++ lib/trilby.nix
@@ -106,13 +106,28 @@ rec {
       };
     };
 
-  trilbyUser = u:
+  trilbyUser = trilby: u:
     let
-      user = {
+      isLinux = trilby.hostSystem.kernel.name == "linux";
+      isDarwin = trilby.hostSystem.kernel.name == "darwin";
+      user =
+      {
         uid = u.uid or 1000;
         name = u.name;
+        home = u.home or (if isDarwin then "/Users/${u.name}" else "/home/${u.name}");
+      }
+      // lib.optionalAttrs isLinux (
+        if (u ? initialPassword && u ? initialHashedPassword)
+        then error "trilbyUser: both `initialPassword` and `initialHashedPassword` cannot be specified"
+        else if (u ? initialPassword)
+        then { inherit (u) initialPassword; }
+        else if (u ? initialHashedPassword)
+        then { inherit (u) initialHashedPassword; }
+        else
+          throw "trilbyUser: required attribute `initialPassword` or `initialHashedPassword` missing"
+      )
+      // lib.optionalAttrs isLinux {
         isNormalUser = u.isNormalUser or true;
-        home = u.home or "/home/${u.name}";
         createHome = u.createHome or true;
         group = u.group or u.name;
         extraGroups = u.extraGroups or [
@@ -121,17 +136,7 @@ rec {
           "video"
           "wheel"
         ];
-      }
-      // (
-        if (u ? initialPassword && u ? initialHashedPassword)
-        then error "trilbyUser: both `initialPassword` and `initialHashedPassword` cannot be specified"
-        else if (u ? initialPassword)
-        then { inherit (u) initialPassword; }
-        else if (u ? initialHashedPassword)
-        then { inherit (u) initialHashedPassword; }
-        else
-          error "trilbyUser: required attribute `initialPassword` or `initialHashedPassword` missing"
-      );
+      };
       group.gid = u.gid or user.uid;
       home = {
         home = {
@@ -144,8 +149,11 @@ rec {
       };
     in
     {
-      users.groups.${user.name} = group;
       users.users.${user.name} = user;
       home-manager.users.${user.name} = home;
+    }
+    //
+    lib.optionalAttrs isLinux {
+      users.groups.${user.name} = group;
     };
 }