+++ modules/profiles/darwin.nix
@@ -1,4 +1,4 @@
-{ inputs, ... }:
+{ inputs, pkgs, ... }:
 
 {
   imports = [
@@ -7,5 +7,19 @@
     inputs.self.nixosModules.profiles.nixpkgs
     inputs.self.nixosModules.profiles.users
     inputs.self.nixosModules.profiles.zsh
+  ];
+
+  environment.systemPackages = with pkgs; [
+    expect
+    file
+    git
+    jq
+    nvd
+    rsync
+    tmux
+    trilby-cli
+    unzip
+    wget
+    zip
   ];
 }