+++ trilby-cli/src/Prelude.hs
@@ -239,7 +239,7 @@ cmd' (p :| args) = do
 cmd :: (HasCallStack) => NonEmpty Text -> App Text
 cmd args = do
     (code, out) <- cmd' args
-    logDebug $ "cmd return code: " <> ishow code
+    logDebug $ "cmd: " <> ishow (code, out)
     case code of
         ExitSuccess -> pure out
         ExitFailure{} -> liftIO $ exitWith code
@@ -304,7 +304,7 @@ shell' cmd stdin = do
 shell :: (HasCallStack) => Text -> Turtle.Shell Turtle.Line -> App Text
 shell cmd stdin = do
     (code, out) <- shell' cmd stdin
-    logDebug $ "shell return code: " <> ishow code
+    logDebug $ "shell: " <> ishow (code, out)
     case code of
         ExitSuccess -> pure out
         ExitFailure{} -> liftIO $ exitWith code