+++ trilby-cli/src/Trilby/Configuration.hs
@@ -3,7 +3,10 @@ module Trilby.Configuration where
 import Trilby.Host
 import Prelude
 
-data Configuration = Configuration {name :: Text, host :: Host}
+data Configuration = Configuration
+    { name :: Text
+    , host :: Host
+    }
     deriving stock (Generic, Eq, Ord, Show)
 
 fromHost :: Host -> App Configuration