+++ trilby-cli/src/Trilby/Host.hs
@@ -3,8 +3,8 @@ module Trilby.Host where
 import Data.List.Extra (split)
 import Data.Text qualified as Text
 import Trilby.App ()
-import Prelude
 import Trilby.System (System)
+import Prelude
 
 data Host
     = Localhost
@@ -54,10 +54,11 @@ reboot r host = whenM r $ ssh host rawCmd_ ["sudo", "systemctl", "reboot"]
 
 hostSystem :: (HasCallStack) => Host -> App System
 hostSystem host = do
-    systemText <- fmap firstLine . ssh host cmd . sconcat $
-        [ ["nix", "eval"]
-        , ["--impure"]
-        , ["--raw"]
-        , ["--expr", "builtins.currentSystem"]
-        ]
+    systemText <-
+        fmap firstLine . ssh host cmd . sconcat $
+            [ ["nix", "eval"]
+            , ["--impure"]
+            , ["--raw"]
+            , ["--expr", "builtins.currentSystem"]
+            ]
     pure . read . Text.unpack $ systemText