+++ trilby-cli/src/Trilby/Install.hs
@@ -1,7 +1,6 @@
 module Trilby.Install (install) where
 
 import Data.Text qualified as Text
-import Trilby.Disko (Disko)
 import Trilby.HNix (FlakeRef (..), writeNixFile)
 import Trilby.Host (Host (Localhost), hostSystem, reboot)
 import Trilby.Install.Config.Host
@@ -10,34 +9,44 @@ import Trilby.Install.Disko
 import Trilby.Install.Disko qualified as Disko
 import Trilby.Install.Flake
 import Trilby.Install.Options
+import Trilby.System
 import Trilby.Widgets
 import Turtle qualified
 import Prelude
 
-rootMount :: Path Abs Dir
-rootMount = $(mkAbsDir "/mnt")
+rootMount :: Kernel -> Path Abs Dir
+rootMount Linux = $(mkAbsDir "/mnt")
+rootMount Darwin = $(mkAbsDir "/")
 
-trilbyDir :: Path Abs Dir
-trilbyDir = trilbyHome rootMount
+trilbyDir :: Kernel -> Path Abs Dir
+trilbyDir = trilbyHome . rootMount
 
 install :: InstallOpts Maybe -> App ()
-install (askOpts -> opts) | Just FlakeOpts{..} <- opts.flake = do
+install (askOpts -> opts) = do
+    system <- hostSystem Localhost
+    case system.kernel of
+        Linux -> installLinux opts
+        Darwin -> installDarwin opts
+
+installLinux :: InstallOpts App -> App ()
+installLinux opts | Just FlakeOpts{..} <- opts.flake = do
     let diskoRef = Disko.Flake flakeRef
     formatDisk opts.format diskoRef
     mountRoot diskoRef
     nixosInstall flakeRef
     whenM copyFlake do
         storePath <- Text.strip <$> shell ("nix flake archive --json " <> flakeRef.url <> " | jq --raw-output .path") empty
-        asRoot cmd_ ["cp", "-r", storePath, fromPath trilbyDir]
-        asRoot cmd_ ["chown", "-R", "1000:1000", fromPath trilbyDir]
+        asRoot cmd_ ["cp", "-r", storePath, fromPath $ trilbyDir Linux]
+        asRoot cmd_ ["chown", "-R", "1000:1000", fromPath $ trilbyDir Linux]
     reboot opts.reboot Localhost
-install (askOpts -> opts) = withTempFile $(mkRelFile "disko.nix") \diskoFile -> do
+installLinux opts = withTempFile $(mkRelFile "disko.nix") \diskoFile -> do
     disko <- getDisko opts
     inDir (parent diskoFile) $ writeNixFile diskoFile disko
     let diskoRef = Disko.File $ Abs diskoFile
     formatDisk opts.format diskoRef
     mountRoot diskoRef
-    flakeRef <- setupHost disko opts
+    flakeRef <- setupHost Linux opts $ \hostDir _ -> do
+        inDir hostDir $ writeNixFile $(mkRelFile "disko.nix") $ sanitise disko
     nixosInstall flakeRef
     reboot opts.reboot Localhost
 
@@ -53,34 +62,57 @@ mountRoot d = unlessM rootIsMounted do
         errorExit "Cannot install without mounted partitions"
     disko $ Mount d
   where
-    rootIsMounted = (ExitSuccess ==) . fst <$> cmd' ["mountpoint", "-q", fromPath rootMount]
+    rootIsMounted = (ExitSuccess ==) . fst <$> cmd' ["mountpoint", "-q", fromPath $ rootMount Linux]
 
 flakeNix, defaultNix, configurationNix :: Path Rel File
 flakeNix = $(mkRelFile "flake.nix")
 defaultNix = $(mkRelFile "default.nix")
 configurationNix = $(mkRelFile "configuration.nix")
 
-setupHost :: Disko -> InstallOpts App -> App FlakeRef
-setupHost disko opts = do
+data Owner = Owner {uid :: Int, gid :: Int}
+
+instance Show Owner where
+    show Owner{..} = show uid <> ":" <> show gid
+
+setupHost
+    :: Kernel
+    -> InstallOpts App
+    -> (Path Rel Dir -> Path Rel Dir -> App ())
+    -> App FlakeRef
+setupHost kernel opts actions = do
     hostname <- opts.hostname
     edition <- opts.edition
     release <- opts.release
-    inDir trilbyDir do
-        asRoot cmd_ ["chown", "-R", "1000:1000", fromPath trilbyDir]
-        writeNixFile flakeNix $ flake release
+    realTrilbyDir <- canonicalizePath $ trilbyDir kernel
+    inDir realTrilbyDir do
+        owner <-
+            case kernel of
+                Linux -> pure Owner{uid = 1000, gid = 1000}
+                Darwin -> do
+                    uid <- read . fromText <$> cmd ["id", "-u"]
+                    gid <- read . fromText <$> cmd ["id", "-g"]
+                    pure Owner{..}
+        asRoot cmd_ ["chown", "-R", ishow owner, fromPath realTrilbyDir]
+        writeNixFile flakeNix $ flake kernel release
+        hostDir <- parseRelDir . fromText $ "hosts/" <> hostname
         username <- opts.username
-        password <- do
-            rawPassword <- opts.password
-            HashedPassword . firstLine <$> Turtle.strict (Turtle.inproc "mkpasswd" [rawPassword] empty)
-        let user = User{uid = 1000, ..}
+        (user, host) <-
+            case kernel of
+                Linux -> do
+                    password <- Just <$> (hashedPassword =<< opts.password)
+                    let user = User{uid = owner.uid, ..}
+                    keyboard <- Just <$> opts.keyboard
+                    locale <- Just <$> opts.locale
+                    timezone <- Just <$> opts.timezone
+                    pure (user, Host{..})
+                Darwin -> do
+                    let user = User{uid = owner.uid, username, password = Nothing}
+                    let host = Host{keyboard = Nothing, locale = Nothing, timezone = Nothing, ..}
+                    pure (user, host)
+
         userDir <- parseRelDir . fromText $ "users/" <> username
         let userFile = userDir </> defaultNix
         writeNixFile userFile user
-        keyboard <- opts.keyboard
-        locale <- opts.locale
-        timezone <- opts.timezone
-        let host = Host{..}
-        hostDir <- parseRelDir . fromText $ "hosts/" <> hostname
         inDir hostDir do
             platform <- show <$> hostSystem Localhost
             writeNixFile
@@ -97,30 +129,35 @@ setupHost disko opts = do
                 }
                 |]
             writeNixFile configurationNix host
-            writeFile $(mkRelFile "hardware-configuration.nix")
-                =<< asRoot
-                    cmd
-                    [ "nixos-generate-config"
-                    , "--show-hardware-config"
-                    , "--no-filesystems"
-                    , "--root"
-                    , fromPath rootMount
-                    ]
-            writeNixFile $(mkRelFile "disko.nix") $ sanitise disko
-        cmd_ . sconcat $
-            [ ["nix", "flake", "lock"]
-            , ["--accept-flake-config"]
-            , ["--override-input", "trilby", "trilby"]
-            ]
+            case kernel of
+                Linux -> do
+                    writeFile $(mkRelFile "hardware-configuration.nix")
+                        =<< asRoot
+                            cmd
+                            [ "nixos-generate-config"
+                            , "--show-hardware-config"
+                            , "--no-filesystems"
+                            , "--root"
+                            , fromPath $ rootMount kernel
+                            ]
+                    cmd_ . sconcat $
+                        [ ["nix", "flake", "lock"]
+                        , ["--accept-flake-config"]
+                        , ["--override-input", "trilby", "trilby"]
+                        ]
+                Darwin -> pure ()
+        actions hostDir userDir
         whenM opts.edit do
-            editor <- getEnv "EDITOR"
+            editor <- fromMaybe "nano" <$> lookupEnv "EDITOR"
             rawCmd_
                 [ fromString editor
                 , "flake.nix"
                 , fromPath $ hostDir </> configurationNix
-                , fromPath $ userDir </> defaultNix
                 ]
-    pure FlakeRef{url = fromPath trilbyDir, output = pure hostname}
+    pure FlakeRef{url = fromPath realTrilbyDir, output = pure hostname}
+
+hashedPassword :: Text -> App Password
+hashedPassword plain = HashedPassword . firstLine <$> Turtle.strict (Turtle.inproc "mkpasswd" [plain] empty)
 
 nixosInstall :: FlakeRef -> App ()
 nixosInstall flakeRef = do
@@ -140,3 +177,10 @@ nixosInstall flakeRef = do
         , ["--no-root-password"]
         , ["--impure"]
         ]
+
+installDarwin :: InstallOpts App -> App ()
+-- installDarwin opts | Just FlakeOpts{..} <- opts.flake = pure ()
+installDarwin opts = do
+    flakeRef <- setupHost Darwin opts \_ _ -> pure ()
+    let darwinRebuild = flakeRef{output = ["darwin-rebuild"]}
+    cmd_ ["nix", "run", ishow darwinRebuild, "--", "switch", "--flake", ishow flakeRef]